move util.enforcetypes to types.apply_annotations
